+++ apps/ex_scim/lib/ex_scim/groups/mapper.ex
@@ -1,11 +1,11 @@
 defmodule ExScim.Groups.Mapper do
-  @spec from_scim(map(), ExScim.Auth.Principal.t()) ::
+  @spec from_scim(map(), ExScim.Scope.t()) ::
           {:ok, struct() | map()} | {:error, atom() | term()}
   def from_scim(scim_data, caller) do
     adapter().from_scim(scim_data, caller)
   end
 
-  @spec to_scim(struct() | map(), ExScim.Auth.Principal.t(), keyword()) ::
+  @spec to_scim(struct() | map(), ExScim.Scope.t(), keyword()) ::
           {:ok, map()} | {:error, atom() | term()}
   def to_scim(group_struct, caller, opts \\ []) do
     adapter().to_scim(group_struct, caller, opts)